Add event linster function
