+++ src/index.html
@@ -3,17 +3,31 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>To Do List </title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-
+    <title>To Do List</title>
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC"
+      crossorigin="anonymous"
+    />
+    <script
+      src="https://kit.fontawesome.com/8107cf6975.js"
+      crossorigin="anonymous"
+    ></script>
   </head>
-  <body> 
+  <body>
     <div class="list-group,m-auto flex align-content-center m-5">
-        <ul id ="list" class="mb-n4">
-            <li class="list-group-item">Today's To Do</li>
-            <li class="list-group-item"><input type="text" class="form-control" id="floatingInput" placeholder=" Add to the List..."></li>
-        </ul>
+      <ul id="list" class="mb-n4">
+        <li class="list-group-item">Today's To Do</li>
+        <li class="list-group-item">
+          <input
+            type="text"
+            class="form-control"
+            id="floatingInput"
+            placeholder=" Add to the List..."
+          />
+        </li>
+      </ul>
     </div>
   </body>
 </html>
-
