+++ src/order.js
@@ -1,0 +1,3 @@
+function reorderTasks(tasks) {
+  return _.orderBy(tasks, ['index'], ['asc']);
+}
